+++ src/main/java/Realtime/inventory/dropCItemButton.java
@@ -37,7 +37,7 @@ public class dropCItemButton extends RenderableButton implements Renderable, Cli
     }
 
     @Override
-    public boolean isInBounds(int pOrX,int pOrY){
+    public boolean isInBounds(double pOrX,double pOrY){
         //If each of the coordinates given (x & y) is within the boundaries of the square trigger. Based on its position and its size.
         return super.isInBounds(pOrX, pOrY);
     }
